Changed ComponentAnnotation to TypeScript
